zenject wip

+++ Assets/_SunsetSystems/Entities/Characters/Data/CreatureData.cs
@@ -24,7 +24,7 @@ namespace SunsetSystems.Entities.Characters
         [ES3Serializable]
         public EquipmentData Equipment;
         public bool UseEquipmentPreset;
-        public float Money;
+        public int Money;
 
         public CreatureData(CreatureConfig config)
         {
